move thread tests to separate file
